+++ src/rgb.rs
@@ -4,31 +4,37 @@
 // Copyright (c) 2019-2020  Jeron Aldaron Lau
 //
 use crate::alpha::{
-    self, Alpha, Mode as _, Opaque, Premultiplied, Straight, Translucent,
+    self, AChannel, Mode as _, Opaque, Premultiplied, Straight, Translucent,
 };
 use crate::gamma::{self, Linear};
 use crate::{Ch16, Ch32, Ch8, Channel, Format};
 use std::marker::PhantomData;
 use std::ops::Mul;
 
-/// RGB color model, with optional [Alpha](alpha/trait.Alpha.html) channel.
+/// RGB color model, with optional [alpha channel](alpha/trait.AChannel.html).
 ///
 /// The `Channel`s are *red*, *green* and *blue*.
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 #[repr(C)]
-pub struct Rgb<C: Channel, A: Alpha, M: alpha::Mode, G: gamma::Mode> {
-    mode: PhantomData<M>,
-    gamma: PhantomData<G>,
+pub struct Rgb<C, A, M, G>
+where
+    C: Channel,
+    A: AChannel<Chan = C>,
+    M: alpha::Mode,
+    G: gamma::Mode,
+{
     red: C,
     green: C,
     blue: C,
     alpha: A,
+    mode: PhantomData<M>,
+    gamma: PhantomData<G>,
 }
 
 impl<C, A, M, G> Iterator for Rgb<C, A, M, G>
 where
     C: Channel,
-    A: Alpha,
+    A: AChannel<Chan = C>,
     M: alpha::Mode,
     G: gamma::Mode,
 {
@@ -64,13 +70,13 @@ where
 impl<C, A, G> From<Rgb<C, A, Straight, G>> for Rgb<C, A, Premultiplied, G>
 where
     C: Channel,
-    A: Alpha<Chan = C>,
+    A: AChannel<Chan = C>,
     G: gamma::Mode,
 {
     fn from(c: Rgb<C, A, Straight, G>) -> Self {
-        let red = Premultiplied::encode::<C, A>(c.red, c.alpha);
-        let green = Premultiplied::encode::<C, A>(c.green, c.alpha);
-        let blue = Premultiplied::encode::<C, A>(c.blue, c.alpha);
+        let red = Premultiplied::encode(c.red, c.alpha());
+        let green = Premultiplied::encode(c.green, c.alpha());
+        let blue = Premultiplied::encode(c.blue, c.alpha());
         Rgb::with_alpha(red, green, blue, c.alpha)
     }
 }
@@ -78,13 +84,13 @@ where
 impl<C, A, G> From<Rgb<C, A, Premultiplied, G>> for Rgb<C, A, Straight, G>
 where
     C: Channel,
-    A: Alpha<Chan = C>,
+    A: AChannel<Chan = C>,
     G: gamma::Mode,
 {
     fn from(c: Rgb<C, A, Premultiplied, G>) -> Self {
-        let red = Premultiplied::decode::<C, A>(c.red, c.alpha);
-        let green = Premultiplied::decode::<C, A>(c.green, c.alpha);
-        let blue = Premultiplied::decode::<C, A>(c.blue, c.alpha);
+        let red = Premultiplied::decode(c.red, c.alpha());
+        let green = Premultiplied::decode(c.green, c.alpha());
+        let blue = Premultiplied::decode(c.blue, c.alpha());
         Rgb::with_alpha(red, green, blue, c.alpha)
     }
 }
@@ -92,7 +98,7 @@ where
 impl<C, A, M, G> From<i32> for Rgb<C, A, M, G>
 where
     C: Channel + From<Ch8>,
-    A: Alpha + From<Translucent<Ch8>>,
+    A: AChannel<Chan = C> + From<Translucent<Ch8>>,
     M: alpha::Mode,
     G: gamma::Mode,
 {
@@ -110,7 +116,7 @@ impl<C, A, M, G> From<Rgb<C, A, M, G>> for i32
 where
     C: Channel,
     Ch8: From<C>,
-    A: Alpha<Chan = C>,
+    A: AChannel<Chan = C>,
     M: alpha::Mode,
     G: gamma::Mode,
 {
@@ -122,7 +128,7 @@ where
         let green = i32::from(green) << 8;
         let blue: u8 = Ch8::from(c.blue()).into();
         let blue = i32::from(blue) << 16;
-        let alpha: u8 = Ch8::from(c.alpha().value()).into();
+        let alpha: u8 = Ch8::from(c.alpha()).into();
         let alpha = i32::from(alpha) << 24;
         red | green | blue | alpha
     }
@@ -131,7 +137,7 @@ where
 impl<C, A, G> Mul<Self> for Rgb<C, A, Straight, G>
 where
     C: Channel,
-    A: Alpha,
+    A: AChannel<Chan = C>,
     G: gamma::Mode,
 {
     type Output = Self;
@@ -154,7 +160,7 @@ where
 impl<C, A, G> Mul<Self> for Rgb<C, A, Premultiplied, G>
 where
     C: Channel,
-    A: Alpha<Chan = C>,
+    A: AChannel<Chan = C>,
     G: gamma::Mode,
 {
     type Output = Self;
@@ -169,7 +175,7 @@ where
 impl<C, A, M, G> Rgb<C, A, M, G>
 where
     C: Channel,
-    A: Alpha,
+    A: AChannel<Chan = C>,
     M: alpha::Mode,
     G: gamma::Mode,
 {
@@ -215,15 +221,15 @@ where
         self.blue
     }
     /// Get the alpha `Channel`.
-    pub fn alpha(self) -> A {
-        self.alpha
+    pub fn alpha(self) -> C {
+        self.alpha.value()
     }
 }
 
 impl<C, A, M, G> Format for Rgb<C, A, M, G>
 where
     C: Channel,
-    A: Alpha<Chan = C> + From<C>,
+    A: AChannel<Chan = C> + From<C>,
     M: alpha::Mode,
     G: gamma::Mode,
 {
@@ -233,7 +239,7 @@ where
 
     /// Get *red*, *green*, *blue* and *alpha* `Channel`s
     fn rgba(self) -> [Self::Chan; 4] {
-        [self.red, self.green, self.blue, self.alpha.value()]
+        [self.red, self.green, self.blue, self.alpha()]
     }
 
     /// Make a pixel with given RGBA `Channel`s
@@ -262,10 +268,10 @@ where
         } else {
             rhs.blue - self.blue
         };
-        let a = if self.alpha.value() > rhs.alpha.value() {
-            self.alpha.value() - rhs.alpha.value()
+        let a = if self.alpha() > rhs.alpha() {
+            self.alpha() - rhs.alpha()
         } else {
-            rhs.alpha.value() - self.alpha.value()
+            rhs.alpha() - self.alpha()
         };
         Rgb::with_alpha(r, g, b, a)
     }
@@ -275,7 +281,7 @@ where
         self.red <= rhs.red
             && self.green <= rhs.green
             && self.blue <= rhs.blue
-            && self.alpha.value() <= rhs.alpha.value()
+            && self.alpha() <= rhs.alpha()
     }
 }